Added Item class and Leaf class inside of KnapsackProblemSolver class.

+++ KnapsackProblemLibrary/Item.cs
@@ -18,7 +18,7 @@ namespace KnapsackProblemLibrary
             Id = id;
             Weight = weight;
             Value = value;
-            UnitValue = (float)value / weight;
+            UnitValue = (double)value / weight;
         }
     }
 }
